Move verify message to utils module
