+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export { generateMnemonic } from './mnemonic'
+export { verifyMessage } from './verify_message'
