+++ src/utils/verify_message.spec.ts
@@ -1,0 +1,42 @@
+import { expect } from 'chai'
+import { MemoryKeyManager } from '../KeyManager'
+import { AddressType } from '../Wallet'
+import { verifyMessage } from './verify_message'
+import { getBindingsForEnvironment } from '../lib/bindings'
+const { AddressKeyIndex } = getBindingsForEnvironment()
+
+describe('utils', () => {
+  describe('verifyMessage', () => {
+    it('returns true when verifying a correct signature for a message', () => {
+      const message = 'foobar'
+      const mnemonic = 'height bubble drama century ask online stage camp point loyal hip awesome'
+      const keyManager = MemoryKeyManager({ mnemonic, password: 'securepassword' })
+      const signatureAsHex = keyManager.signMessage(AddressType.external, 0, message)
+      const publicKey = keyManager.publicAccount().address_key(false, AddressKeyIndex.new(0))
+
+      const verification = verifyMessage({
+        publicKey,
+        message,
+        signatureAsHex
+      })
+
+      expect(verification).to.eql(true)
+    })
+
+    it('returns false when verifying an incorrect message for a valid signature', () => {
+      const message = 'foobar'
+      const mnemonic = 'height bubble drama century ask online stage camp point loyal hip awesome'
+      const keyManager = MemoryKeyManager({ mnemonic, password: 'securepassword' })
+      const signatureAsHex = keyManager.signMessage(AddressType.external, 0, message)
+      const publicKey = keyManager.publicAccount().address_key(false, AddressKeyIndex.new(0))
+
+      const verification = verifyMessage({
+        publicKey,
+        message: 'a differnt message',
+        signatureAsHex
+      })
+
+      expect(verification).to.eql(false)
+    })
+  })
+})
